requirements.txt
Package              Version  
-------------------- ---------
aiohttp              3.5.4    
aiomysql             0.0.20   
Appium-Python-Client 0.31     
asn1crypto           0.24.0   
astroid              2.0      
async-timeout        3.0.1    
attrs                18.1.0   
Automat              0.7.0    
bcrypt               3.1.7    
beautifulsoup4       4.7.1    
blinker              1.4      
brotlipy             0.7.0    
bs4                  0.0.1    
certifi              2018.4.16
cffi                 1.11.5   
chardet              3.0.4    
click                6.7      
colorama             0.3.9    
constantly           15.1.0   
cryptography         2.7      
cssselect            1.0.3    
cycler               0.10.0   
et-xmlfile           1.0.1    
Flask                1.0.3    
h11                  0.7.0    
h2                   3.1.0    
hpack                3.0.0    
hyperframe           5.2.0    
hyperlink            18.0.0   
idna                 2.7      
incremental          17.5.0   
isort                4.3.4    
itchat               1.3.10   
itsdangerous         1.1.0    
jdcal                1.4      
Jinja2               2.10.1   
kaitaistruct         0.8      
kiwisolver           1.0.1    
lazy-object-proxy    1.3.1    
ldap3                2.5.2    
lxml                 4.2.4    
MarkupSafe           1.1.1    
matplotlib           3.0.0    
mccabe               0.6.1    
mitmproxy            4.0.4    
multidict            4.5.2    
numpy                1.15.1   
opencv-python        3.4.2.17 
openpyxl             2.5.9    
pandas               0.24.2   
paramiko             2.6.0    
parsel               1.5.0    
passlib              1.7.1    
pexpect              4.6.0    
Pillow               5.2.0    
pip                  19.1     
psutil               5.4.7    
ptyprocess           0.6.0    
pyasn1               0.4.4    
pyasn1-modules       0.2.2    
PyAutoGUI            0.9.38   
pycparser            2.18     
pycryptodome         3.6.5    
pycurl               7.43.0.2 
pyDes                2.0.1    
pydgutils            0.0.16   
PyDispatcher         2.0.5    
pydivert             2.1.0    
PyExecJS             1.5.1    
PyHamcrest           1.9.0    
pylint               2.0.0    
PyMsgBox             1.0.6    
PyMySQL              0.9.2    
PyNaCl               1.3.0    
pyOpenSSL            18.0.0   
pyparsing            2.2.1    
pyperclip            1.6.5    
pypng                0.0.18   
PyQRCode             1.2.1    
PyQt5                5.11.3   
PyQt5-sip            4.19.13  
PyScreeze            0.1.18   
PySocks              1.6.8    
pytesseract          0.2.4    
python-dateutil      2.7.3    
PyTweening           1.0.3    
pytz                 2019.1   
pywin32              223      
pywinio              0.3.0    
queuelib             1.5.0    
rabird.core          0.4.0    
rabird.winio         0.2.6    
requests             2.21.0   
ruamel.yaml          0.15.94  
scipy                1.1.0    
Scrapy               1.5.1    
selenium             3.141.0  
service-identity     17.0.0   
setuptools           39.0.1   
six                  1.11.0   
sortedcontainers     2.0.5    
soupsieve            1.9.1    
SQLAlchemy           1.2.11   
tornado              5.1.1    
Twisted              18.7.0   
urllib3              1.23     
urwid                2.0.1    
w3lib                1.19.0   
Werkzeug             0.15.4   
winiobinary          0.0.6    
wrapt                1.10.11  
wsproto              0.11.0   
xlrd                 1.1.0    
xlutils              2.0.0    
xlwt                 1.3.0    
yarl                 1.3.0    
zope.interface       4.5.0    
